Refactor default_keymaps, fix config default, add test
